refactor: add updatePackeg file
